S104 split: extract instance_setup.py (instance + DC + polaris)
main.py: 3,131 -> 2,293 lines. main.py keeps explicit re-exports
(from x import y as y) for BootstrapPlan, DCSetupPlan,
build_guest_execution_context, SetupOrchestrator, OpsOrchestrator,
_should_promote_dc_at_runtime, _should_run_dc_bootstrap_plan so the
sub-module's lazy main.X(...) lookups resolve at runtime, mypy
treats them as explicit re-exports, and the existing
patch("main.X") test mocks intercept the same call sites.
